Accept Import: Add webservice: Accept Scan Filename with Path
Calling programs can now call a new webservice to pass a Scan filename
and it's path.
This would be called in place of sending the scan file.
The Spectral Storage Service webapp and importer need to be able to read
the file as sent, which is checked in the webapp during the webservice
call.
New Configuration value is required to be configured to support this new
feature.