Accept Import: Process Import code
Updated/Fixed the code for getting the next scan file to import and
executing the Importer for it.
The previous code did get the processing done correctly but was
incorrect in where it waited.
The previous code didn't process all available scan files and thus if it
was processing a scan file when another was submitted, the submitted
scan file was not processed until the extra long wait() [Extra long
since never should find something to process when that wait exits on
timeout] exitted.
This has all been fixed.