Update Background Thread processing, especially on webapp undeploy
Update Background Thread processing on shutdown to ensure the threads
exit when the webapp undeploys.
Added an extra flag and check that when the undeploy starts that the
background threads will for sure not start any additional processing and
will exit existing processing quickly and reset the status of any in
progress imports.  Changes can be done to improve the importer process
(Done for some areas but others have been identified) to also help this
and will be scheduled for later work.