Importer (Accept Import config): Config option Dir for temp write files.
For installations where the main directory that Spectr writes it's data
files to is being backed up using a diffing tool (like rdiff), writing
the files first to the temp directory under that directory may generate
additional differences even though they are in that directory only for a
short while.
This new config option allows those files to be first written to a
directory that is not backed up and then moved under the main storage
directory.