Comment out all AWS S3 support since has not been maintained or tested
The AWS S3 support has not been maintained or tested as new features
have been added so comment them out and error if configured to use S3.

S3 can be used in the future but all functionality through the code
would need to be reviewed for S3 support and not just uncomment the S3
code.