Get Data:S File Core:Put In Memory Data Cache inside Java SoftReference
Put uses of Google Cache inside a Java SoftReference to provide a way
for Java garbage collector to drop the whole in memory cache when out of
heap memory to avoid getting out of heap memory error.

Ideally the cache would be sized to fit in configured heap memory but
using SoftReference provides extra protection.