Java Client Connection Jar build changes
For the client connection jars provided to provide a way to call the
webservices from Java, the approach to building them has changed.

Gradle 'shadowJar' plugin is no longer used.
They are created using standard Gradle build jar and then Ant is used to
combine the Common code jar with the Client Connector jar into a single
jar for deployment for each of "Accept Import" and "Get Data"