Get Data Webapp: Add Return Scan Peak with Max Intensity (Separate field
Only Return Scan Peak with Max Intensity when Requested to support
connection from older client.

This was primarily added to get the Scan Peak with Max Intensity in the
scan when filtering scan peaks.

The Scan Peak with Max Intensity is returned in a separate property from
the Scan Peak List.
So it is possible that the Scan Peak with Max Intensity is returned
twice, once in the list and once in the separate property.