Web: Get Data: Get Scan Data: Return Scans in order requested
For when NO Parent scans are requested, the scans will be returned in
the order the scan numbers are in the request.
This was originally the case.
The order the scans were returned became indeterminate when the scan
data retrieval was run through Java 'parallelStream()'.
Now all the retrieved scans are put in a Map and then the original scan
numbers are processed so that the scans are returned in scan numbers
request order so back to the original result.